107 create endpoint tests (#132)
Tests for server_backend created
