+++ .gitignore
@@ -34,6 +34,9 @@ venv/
 
 config.json
 
+# Pytest stuff
+.coverage
+
 # Passwords for database
 .env
 # URL / Port for Server Flask. Every developer should have a different port for development. Production wll be 1 port.
