+++ server_backend/requirements.txt
@@ -1,19 +1,36 @@
 blinker==1.9.0
 click==8.1.7
+coverage==7.6.12
+exceptiongroup==1.2.2
 Flask==3.0.3
 Flask-Cors==5.0.0
 Flask-MySQLdb==2.0.0
 gunicorn==23.0.0
+iniconfig==2.0.0
 itsdangerous==2.2.0
 Jinja2==3.1.4
 MarkupSafe==3.0.2
+mirakuru==2.6.0
+mypy-extensions==1.0.0
+mysql-connector-python==9.2.0
 mysqlclient==2.2.6
 numpy==2.2.0
 packaging==24.2
 pandas==2.2.3
+pathspec==0.12.1
+platformdirs==4.3.6
+pluggy==1.5.0
+port-for==0.7.4
+psutil==7.0.0
+PyMySQL==1.1.1
+pytest==8.3.4
+pytest-cov==6.0.0
+pytest-mysql==3.1.0
 python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
 pytz==2024.2
 six==1.17.0
+tomli==2.2.1
+typing_extensions==4.12.2
 tzdata==2024.2
 Werkzeug==3.1.3
